+++ src/plugins/replication/replication-helper.ts
@@ -3,6 +3,7 @@ import type {
     WithDeleted
 } from '../../types';
 import { flatClone } from '../../plugins/utils';
+import { getComposedPrimaryKeyOfDocumentData } from '../../rx-schema-helper';
 
 // does nothing
 export const DEFAULT_MODIFIER = (d: any) => Promise.resolve(d);
@@ -23,20 +24,41 @@ export function swapDefaultDeletedTodeletedField<RxDocType>(
     }
 }
 
-
-export function swapdeletedFieldToDefaultDeleted<RxDocType>(
+/**
+ * Must be run over all plain document data
+ * that was pulled from the remote.
+ * Used to fill up fields or modify the deleted field etc.
+ */
+export function handlePulledDocuments<RxDocType>(
+    collection: RxCollection<RxDocType>,
     deletedField: string,
-    doc: RxDocType
-): WithDeleted<RxDocType> {
-    if (deletedField === '_deleted') {
-        return doc as any;
-    } else {
-        doc = flatClone(doc);
-        const isDeleted = !!(doc as any)[deletedField];
-        (doc as any)._deleted = isDeleted;
-        delete (doc as any)[deletedField];
-        return doc as any;
-    }
+    docs: RxDocType[]
+): WithDeleted<RxDocType>[] {
+    return docs.map(doc => {
+        const useDoc: WithDeleted<RxDocType> = flatClone(doc) as any;
+
+        /**
+         * Swap out the deleted field
+         */
+        if (deletedField !== '_deleted') {
+            const isDeleted = !!(useDoc as any)[deletedField];
+            (useDoc as any)._deleted = isDeleted;
+            delete (useDoc as any)[deletedField];
+        } else {
+            // ensure we have a boolean.
+            useDoc._deleted = !!useDoc._deleted;
+        }
+
+        /**
+         * Fill up composed primary
+         */
+        const primaryPath = collection.schema.primaryPath;
+        (useDoc as any)[primaryPath] = getComposedPrimaryKeyOfDocumentData(
+            collection.schema.jsonSchema,
+            useDoc
+        );
+        return useDoc as any;
+    });
 }